Move SettingsView components into parent folder.
