+++ assets/js/modules/analytics/components/settings/GA4SettingsView.js
@@ -26,15 +26,15 @@ import { createInterpolateElement, Fragment } from '@wordpress/element';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import DisplaySetting from '../../../../../components/DisplaySetting';
-import { MODULES_ANALYTICS } from '../../../datastore/constants';
+import DisplaySetting from '../../../../components/DisplaySetting';
+import { MODULES_ANALYTICS } from '../../datastore/constants';
 import {
 	MODULES_ANALYTICS_4,
 	PROPERTY_CREATE,
-} from '../../../../analytics-4/datastore/constants';
-import Link from '../../../../../components/Link';
-import VisuallyHidden from '../../../../../components/VisuallyHidden';
-import { escapeURI } from '../../../../../util/escape-uri';
+} from '../../../analytics-4/datastore/constants';
+import Link from '../../../../components/Link';
+import VisuallyHidden from '../../../../components/VisuallyHidden';
+import { escapeURI } from '../../../../util/escape-uri';
 const { useSelect } = Data;
 
 export default function GA4SettingsView() {
