+++ assets/js/modules/analytics/components/settings/UASnippetSettingsView.js
@@ -25,7 +25,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import { MODULES_ANALYTICS } from '../../../datastore/constants';
+import { MODULES_ANALYTICS } from '../../datastore/constants';
 const { useSelect } = Data;
 
 export default function UASnippetSettingsView() {
